move TextComponent to Beta::Text
