+++ docs/content/components/iconbutton.md
@@ -48,7 +48,7 @@ if your `IconButton` renders a magnifying glass icon and invokves a search actio
 
 <%= render(Primer::BorderBoxComponent.new) do |component| %>
   <% component.body do %>
-    <%= render(Primer::TextComponent.new(pr: 2)) { "Body" } %>
+    <%= render(Primer::Beta::Text.new(pr: 2)) { "Body" } %>
     <%= render(Primer::IconButton.new(icon: :pencil, box: true, "aria-label": "Edit")) %>
   <% end %>
 <% end %>
