+++ docs/content/components/navigationtab.md
@@ -112,7 +112,7 @@ The Tab's text.
 
 | Name | Type | Default | Description |
 | :- | :- | :- | :- |
-| `kwargs` | `Hash` | N/A | The same arguments as [Text](/components/text). |
+| `kwargs` | `Hash` | N/A | The same arguments as [BetaText](/components/betatext). |
 
 ### `Counter`
