+++ static/statuses.json
@@ -6,6 +6,7 @@
   "Primer::AvatarStackComponent": "beta",
   "Primer::BaseButton": "beta",
   "Primer::BaseComponent": "beta",
+  "Primer::Beta::Text": "beta",
   "Primer::BlankslateComponent": "beta",
   "Primer::BorderBoxComponent": "beta",
   "Primer::BoxComponent": "stable",
@@ -44,7 +45,6 @@
   "Primer::SubheadComponent": "beta",
   "Primer::TabContainerComponent": "alpha",
   "Primer::TabNavComponent": "beta",
-  "Primer::TextComponent": "beta",
   "Primer::TimeAgoComponent": "beta",
   "Primer::TimelineItemComponent": "beta",
   "Primer::TimelineItemComponent::BadgeComponent": "alpha",
